Move debug functions to cli/debug package
