+++ cli/debug/debug.go
@@ -1,4 +1,4 @@
-package utils
+package debug
 
 import (
 	"os"
@@ -6,21 +6,21 @@ import (
 	"github.com/Sirupsen/logrus"
 )
 
-// EnableDebug sets the DEBUG env var to true
+// Enable sets the DEBUG env var to true
 // and makes the logger to log at debug level.
-func EnableDebug() {
+func Enable() {
 	os.Setenv("DEBUG", "1")
 	logrus.SetLevel(logrus.DebugLevel)
 }
 
-// DisableDebug sets the DEBUG env var to false
+// Disable sets the DEBUG env var to false
 // and makes the logger to log at info level.
-func DisableDebug() {
+func Disable() {
 	os.Setenv("DEBUG", "")
 	logrus.SetLevel(logrus.InfoLevel)
 }
 
-// IsDebugEnabled checks whether the debug flag is set or not.
-func IsDebugEnabled() bool {
+// IsEnabled checks whether the debug flag is set or not.
+func IsEnabled() bool {
 	return os.Getenv("DEBUG") != ""
 }
